ml-service/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
flask-cors
pytesseract
Pillow
